Modified structure tables

+++ CreateStructure.sql
@@ -8,7 +8,7 @@ CREATE TABLE dbo.SKU(
 CREATE TABLE dbo.Family(
     ID INT IDENTITY(1, 1),
     SurName VARCHAR(100),
-    BudgetValue DECIMAL(9, 2),
+    BudgetValue DECIMAL(18, 2),
     CONSTRAINT PK_Family_ID PRIMARY KEY (ID)
 );
 
@@ -17,9 +17,9 @@ CREATE TABLE dbo.Basket(
     ID_SKU INT NOT NULL,
     ID_Family INT NOT NULL,
     Quantity INT,
-    Value INT,
+    Value DECIMAL(18, 2),
     PurchaseDate DATE DEFAULT(SYSDATETIME()),
-    DiscountValue INT,
+    DiscountValue DECIMAL(18, 2),
     CONSTRAINT PK_Basket_ID PRIMARY KEY (ID),
     CONSTRAINT FK_Basket_ID_SKU FOREIGN KEY (ID_SKU) REFERENCES dbo.SKU(PK_SKU_ID) ON DELETE CASCADE,
     CONSTRAINT FK_Family_ID_SKU FOREIGN KEY (ID_Family) REFERENCES dbo.Family(PK_Family_ID) ON DELETE CASCADE,
